requirements.txt
coloredlogs
iMail
matplotlib
numpy
pandas
Pillow
PyYAML
scikit_learn
scipy
# torch_scatter
# torch==2.4.0+cu118
# torchvision==0.19.0+cu118
tqdm
huggingface-hub
open_clip_torch
psutil
requests
tokenizers
transformers
urllib3